remove extra fixture
