+++ test/Microsoft.Azure.WebJobs.Extensions.Sql.Tests.csproj
@@ -3,9 +3,7 @@
     <PackageReference Include="Microsoft.Azure.Functions.Extensions" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" />
     <PackageReference Include="Moq" />
-    <PackageReference Include="xunit.assert" />
-    <PackageReference Include="xunit.core" />
-    <PackageReference Include="xunit.analyzers" />
+    <PackageReference Include="xunit" />
     <PackageReference Include="xunit.runner.visualstudio" />
     <PackageReference Include="Microsoft.NET.Sdk.Functions" />
     <PackageReference Include="Newtonsoft.Json" />
